+++ test/json.test.js
@@ -181,7 +181,9 @@ describe('toJsonSchema', () => {
         {
           type: 'string',
         },
-        { type: 'number' },
+        {
+          type: 'number',
+        },
       ],
     });
   });
@@ -439,6 +441,47 @@ describe('toJsonSchema', () => {
       },
       required: [],
       additionalProperties: false,
+    });
+  });
+
+  it('should strip custom extensions', async () => {
+    const schema = yd.object({
+      name: yd.string().tag({
+        'x-schema': 'my-schema',
+      }),
+    });
+    const result = schema.toJsonSchema({
+      stripExtensions: true,
+    });
+    expect(result).toEqual({
+      type: 'object',
+      properties: {
+        name: {
+          type: 'string',
+        },
+      },
+      required: [],
+      additionalProperties: false,
+    });
+  });
+
+  it('should strip custom extensions on alternates', async () => {
+    const schema = yd.allow(
+      yd.string().tag({ 'x-name': 'string' }),
+      yd.number().tag({ 'x-name': 'number' }),
+    );
+    const result = schema.toJsonSchema({
+      stripExtensions: true,
+    });
+    expect(result).toEqual({
+      anyOf: [
+        {
+          type: 'string',
+        },
+        {
+          type: 'number',
+        },
+      ],
     });
   });
 });